tests/wpt/web-platform-tests/app-history/currentchange-event/event-constructor.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_js(TypeError, () => {
    new AppHistoryCurrentChangeEvent("currentchange");
  });
}, "can't bypass required members by omitting the dictionary entirely");

test(() => {
  assert_throws_js(TypeError, () => {
    new AppHistoryCurrentChangeEvent("currentchange", {
      navigationType: "push"
    });
  });
}, "from is required");

test(() => {
  const event = new AppHistoryCurrentChangeEvent("currentchange", {
    navigationType: "replace",
    from: appHistory.current
  });
  assert_equals(event.navigationType, "replace");
  assert_equals(event.from, appHistory.current);
}, "all properties are reflected back");

test(t => {
  const event = new AppHistoryCurrentChangeEvent("currentchange", { from: appHistory.current });
  assert_equals(event.navigationType, null);
}, "defaults are as expected");
</script>
